pyproject.toml
[project]
name = "voiceflow-dispatcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-cognitiveservices-speech>=1.47.0",
    "pyaudio>=0.2.14",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "rich>=14.2.0",
    "openai>=1.50.0",
    "azure-servicebus>=7.12.0",
    "aiohttp>=3.9.0",
]
